+++ www/assets/mcpReleaseStore-G3seIQXK.js
@@ -1,0 +1,1 @@
+import{d as t}from"./react-CMd_R4qx.js";const r=t(s=>({flags:{},setAll:e=>s({flags:e}),set:(e,a)=>s(l=>({flags:{...l.flags,[e]:a}}))}));export{r as u};